requirements_render.txt
discord.py==2.5.2
flask==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.41